Code tune

+++ Kotlin/lib/src/main/java/com/wolt/blurhashkt/Base83.kt
@@ -10,11 +10,11 @@ internal object Base83 {
             '-', '.', ':', ';', '=', '?', '@', '[', ']', '^', '_', '{', '|', '}', '~'
     )
 
-	fun encode(value: Int, length: Int, buffer: CharArray, offset: Int) {
+    fun encode(value: Int, length: Int, buffer: CharArray, offset: Int) {
         var exp = 1
         var i = 1
         while (i <= length) {
-            val digit = (value / exp % 83).toInt()
+            val digit = (value / exp % 83)
             buffer[offset + length - i] = CHAR_MAP[digit]
             i++
             exp *= 83
